accounts: move budget test
